refactor(cart): improve Livewire cart trait structure and fix currency formatting
- Remove unused cache blade component
- Change currency helper to use floatval instead of intval for proper decimal handling
- Simplify alert trait by using named parameters in dispatch calls
- Refactor LivewireCartTrait to use computed properties and improve initialization flow
- Add quotation detection for cart instance type
- Remove deprecated property getters in favor of computed properties

+++ app/Helpers/helpers.php
@@ -37,7 +37,7 @@ if ( ! function_exists('format_currency')) {
 
         // $value change to type int|float, string given,
         return Number::currency(
-            intval($value),
+            floatval($value),
             $currencyCode,
             $locale
         );
